ストアを別ファイルに移動
